+++ HTTPCore/trunk/Scanner/RouterAuth.cpp
@@ -66,7 +66,7 @@ static int BruteforceAuth( HTTPAPI *api,HTTPHANDLE HTTPHandle,PREQUEST data,stru
 			if (!CookieNeeded)
 			{
 				//new_response=api->SendHttpRequest( HTTPHandle,NULL,AuthData->method,AuthData->authurl,AuthData->postdata,(unsigned int)strlen(AuthData->postdata),userpass[k].UserName,userpass[k].Password,challenge);
-				new_response=api->SendHttpRequest( HTTPHandle,NULL,AuthData->method,AuthData->authurl,AuthData->postdata,(unsigned int)strlen(AuthData->postdata),userpass[k].UserName,userpass[k].Password);
+				new_response=api->SendHttpRequest( HTTPHandle,NULL,AuthData->method,AuthData->authurl,AuthData->postdata,strlen(AuthData->postdata),userpass[k].UserName,userpass[k].Password);
 			} else {
 				api->SetHTTPConfig(HTTPHandle,ConfigCookie,AuthData->postdata);
 				//new_response=api->SendHttpRequest( HTTPHandle,NULL,AuthData->method,AuthData->authurl,NULL,0,userpass[k].UserName,userpass[k].Password,challenge);
@@ -169,7 +169,7 @@ PREQUEST CheckRouterAuth(HTTPAPI *api,HTTPHANDLE HTTPHandle,PREQUEST data,int nR
 						api->SetHTTPConfig(HTTPHandle,ConfigCookie,(const char*)NULL);
 						free(lpcookie);
 					} else {
-						response=api->SendHttpRequest( HTTPHandle,NULL,AuthData[i].method,AuthData[i].authurl,AuthData[i].postdata,(unsigned int)strlen(AuthData[i].postdata),NULL,NULL);
+						response=api->SendHttpRequest( HTTPHandle,NULL,AuthData[i].method,AuthData[i].authurl,AuthData[i].postdata,strlen(AuthData[i].postdata),NULL,NULL);
 					}
 				}
 				//SetHTTPConfig(HTTPHandle,OPT_HTTP_COOKIE,NULL);
